Organizing, tuning many parameters, adding SAHC strategy, and fixing the sim annealing.
organizing the code more and tuning it.
tuning the heuristic parameters more to make it more informed.
tuning the sim annealing parameters and fixing a typo that caused a major bug, and reworking the upgrades that I've done to the original algo.
adding SAHC (steepest_ascent_hill_climbing) as a fast strategy.
reworking the neighbor getting method - now you can control the operations that each algorithm can use.