Adding support for parallel courses. Tuning and finalizing.
now, if the program adds one course it has to take its parallels (if they exist), and if it removes one course, it removes the courses that have the removed one as a parallel.
also, fixing typos and such, and polishing the code for final push.
added a goal to the next time in Features_to_come.out - it is shown inside it - as I didn't originally think of using libraries to support c/c++ like pointers.